refactor(frontend): move the Old Maid-only types out of common.ts
Review caught the one place where the split broke its own rule. common.ts
says "nothing here belongs to one game", but CpuAction and DrawHistoryEntry
were sitting in it while being used only by Old Maid -- verified: their only
consumers are oldmaid.ts, useOldMaidGame.ts, OldMaidPage.tsx and
OldMaidDrawHistory.tsx. Moved to types/games/oldmaid.ts, which no longer
imports them from ../common.

Beyond the inconsistency, a bare `CpuAction` in the shared namespace is a
latent collision: any other game wanting that name would have had to work
around it, which is exactly the footgun the per-game prefixes avoid
everywhere else.

common.ts is now 10 declarations, all genuinely cross-game.

Also reconciles the count the review raised. card.ts held 894 *type*
declarations (824 interface + 70 type, matching the issue) plus one runtime
export, isMaskedCard -- 895 exported declarations in total. common.ts (10) +
games/*.ts (885) = 895, so nothing was dropped or duplicated. My earlier
"894" counted only interface and type.
