+++ frontend/src/types/games/oldmaid.ts
@@ -1,7 +1,7 @@
 // Type declarations for oldmaid. Split out of card.ts (issue #4366);
 // card.ts re-exports this file, so existing imports keep working.
 
-import type { BaseGameResponse, Card, CpuAction, DrawHistoryEntry } from '../common';
+import type { BaseGameResponse, Card } from '../common';
 
 /** Exported old maid human profile data. */
 export interface OldMaidHumanProfileData {
@@ -48,4 +48,23 @@ export interface OldMaidResponse extends BaseGameResponse {
   mode: number;
   metaAI?: OldMaidMetaAI;
   profile?: OldMaidHumanProfileData;
+}
+
+/** CPU draw/discard action in Old Maid. */
+export interface CpuAction {
+  drawPlayerIdx: number;
+  drawFromIdx: number;
+  drawnCard: Card | null;
+  discardedPairs: number;
+  discardedCards?: Card[];
+  hesitationMs?: number;
+}
+
+/** History entry for a card draw in Old Maid. */
+export interface DrawHistoryEntry {
+  drawPlayerIdx: number;
+  drawFromIdx: number;
+  discardedPairs: number;
+  drawerFinished: boolean;
+  targetFinished: boolean;
 }
